requirements.txt
requests==2.32.3
python-telegram-bot==13.15
dotenv==0.9.9
python-dotenv==1.0.1